# carritodecompras/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Shopping Cart API
  description: API for managing a shopping cart system including products, users, purchases, and detailed purchase history.
  version: v1.0.0
  contact:
    name: Alejandro Granados
    email: alejandro.magb@gmail.com
    url: https://github.com/aledelorbe
...
paths:
...
  /api/products:
    post:
      tags:
        - Products
      summary: Create a product
      description: Creates a new product.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrors'
  /api/products/public/{id}:
    get:
      tags:
        - Products
      summary: Get public product details
      description: Returns public properties of a product by ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The product ID
      responses:
        '200':
          description: Product details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductUserResponseProjection'
        '404':
          description: Product not found
...
  /api/purchases/{userId}/purchases:
    get:
      tags:
        - Purchases
      summary: Get purchases by User ID
      description: Returns the history of purchases for a specific user.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user
      responses:
        '200':
          description: Purchase history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PurchaseHistory'
        '404':
          description: User not found
  /api/purchases/{userId}/purchase:
    post:
      tags:
        - Purchases
      summary: Make a purchase
      description: Creates a new purchase for the user from items in their shopping cart. Adjusts product stocks.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CartItemRequestDto'
      responses:
        '201':
          description: Purchase completed successfully. Returns the updated User object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation errors (e.g. out of stock, empty cart, invalid quantity)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrors'
        '404':
          description: User or Product not found
  /api/purchases/{userId}/purchases/{purchaseId}:
    get:
      tags:
        - Purchases
      summary: Get purchase details
      description: Returns the items and details of a specific purchase for a user.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user
        - name: purchaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the purchase record
      responses:
        '200':
          description: Detailed purchase records retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DetailedPurchaseHistoryDto'
        '404':
          description: User not found

  # Detail controller -----------------------------
  /api/details:
    post:
      tags:
        - Details
      summary: Add a detailed purchase item (Test endpoint)
      description: Creates a single detailed purchase record. Used for testing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItemRequestDto'
      responses:
        '201':
          description: Detailed purchase history created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedPurchaseHistory'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrors'
  /api/details/many:
    post:
      tags:
        - Details
      summary: Add multiple detailed purchase items (Test endpoint)
      description: Creates multiple detailed purchase records. Used for testing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CartItemRequestDto'
      responses:
        '201':
          description: Detailed purchase histories created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DetailedPurchaseHistory'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrors'
components:
  schemas:
    Product:
      type: object
      required:
        - name
        - description
        - brand
        - category
        - price
        - stock
        - image
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Wireless Headphones"
        description:
          type: string
          example: "Active noise-cancelling Bluetooth headphones."
        brand:
          type: string
          example: "Sony"
        category:
          type: string
          example: "Electronics"
        price:
          type: number
          format: double
          example: 199.99
        stock:
          type: integer
          format: int64
          example: 50
        image:
          type: string
          example: "https://example.com/images/headphones.jpg"
        status:
          type: integer
          format: int64
          example: 1
    ProductUserResponseProjection:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Wireless Headphones"
        description:
          type: string
          example: "Active noise-cancelling Bluetooth headphones."
        brand:
          type: string
          example: "Sony"
        category:
          type: string
          example: "Electronics"
        price:
          type: number
          format: double
          example: 199.99
        image:
          type: string
          example: "https://example.com/images/headphones.jpg"
...
    User:
      type: object
      required:
        - name
        - lastname
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John"
        lastname:
          type: string
          example: "Doe"
        purchases:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseHistory'
...
    PurchaseHistory:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        firstImage:
          type: string
          example: "https://example.com/images/headphones.jpg"
        total:
          type: number
          format: double
          example: 399.98
        date:
          type: string
          format: date
          example: "2026-06-20"
    DetailedPurchaseHistory:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int64
          example: 2
        total:
          type: number
          format: double
          example: 399.98
    DetailedPurchaseHistoryDto:
      type: object
      properties:
        detail:
          $ref: '#/components/schemas/DetailedPurchaseHistory'
        name:
          type: string
          example: "Wireless Headphones"
        image:
          type: string
          example: "https://example.com/images/headphones.jpg"
    CartItemRequestDto:
      type: object
      required:
        - idProduct
        - quantity
      properties:
        idProduct:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          format: int64
          example: 2
...
    ValidationErrors:
      type: object
      description: Validation error messages indexed by field name.
      additionalProperties:
        type: string
      example:
        name: "The field name must be obligated"
...